Move marker targeting tests
