+++ helpers/__init__.py
@@ -4,7 +4,7 @@ try:
 except ModuleNotFoundError:
     bpy = None
 
-from .feature_math import (
+from .marker_targeting import (
     calculate_base_values,
     apply_threshold_to_margin_and_distance,
     marker_target_aggressive,
